Added new constant and file
Crearte app.js and added constants SCHEDULE_TIME
